HOTEL
Hotel reservation site. This site is a project that includes features such as countries and cities with agreements, suggestions for places to visit in those cities, hotels with agreements, check-in dates and check-out dates, personal information of the person making the reservation, information on how many adults and children will stay, payment information, payment options (cash-credit card).